# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "summary": "Login user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token successfully refreshed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh JWT token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User successfully logged out."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/users/me": {
      "get": {
        "operationId": "getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/warehouses/{warehouseId}/goods": {
      "get": {
        "operationId": "findAllGoods",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for goods",
            "schema": {
              "example": "Mouse",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goods retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGoodResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List goods for a warehouse",
        "tags": [
          "Goods"
        ]
      },
      "post": {
        "operationId": "createGood",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Good created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new good in a warehouse",
        "tags": [
          "Goods"
        ]
      }
    },
    "/warehouses/{warehouseId}/goods/{id}": {
      "get": {
        "operationId": "findOneGood",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Good retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Good not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific good by id",
        "tags": [
          "Goods"
        ]
      },
      "patch": {
        "operationId": "updateGood",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Good updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Good not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a specific good by id",
        "tags": [
          "Goods"
        ]
      },
      "delete": {
        "operationId": "removeGood",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Good deleted successfully."
          },
          "404": {
            "description": "Good not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a specific good by id",
        "tags": [
          "Goods"
        ]
      }
    }
  },
  "info": {
    "title": "API Documentation",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "example": "StrongP@ss1!",
            "description": "The password of the user"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "The first name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "The last name of the user"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT Access Token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT Refresh Token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "example": "StrongP@ss1!",
            "description": "The password of the user"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "abcde12345...",
            "description": "The refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the user"
          },
          "email": {
            "type": "string",
            "description": "The email of the user"
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the user"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the user"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The creation date of the user record"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The last update date of the user record"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "Updated first name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Updated last name"
          }
        }
      },
...
      "GoodResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the good"
          },
          "name": {
            "type": "string",
            "description": "The name of the good"
          },
          "description": {
            "type": "object",
            "description": "The description of the good",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "description": "The quantity of the good"
          },
          "price": {
            "type": "number",
            "description": "The price of the good"
          },
          "warehouseId": {
            "type": "string",
            "description": "The warehouse ID"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The last update date"
          }
        },
        "required": [
          "id",
          "name",
          "quantity",
          "price",
          "warehouseId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedGoodMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "totalItems",
          "totalPages"
        ]
      },
      "PaginatedGoodResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoodResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedGoodMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateGoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Wireless Mouse",
            "description": "The name of the good"
          },
          "description": {
            "type": "string",
            "example": "A highly responsive wireless mouse.",
            "description": "The description of the good"
          },
          "quantity": {
            "type": "number",
            "example": 100,
            "description": "The quantity of the good available"
          },
          "price": {
            "type": "number",
            "example": 25.99,
            "description": "The price of the good"
          }
        },
        "required": [
          "name",
          "quantity",
          "price"
        ]
      },
      "UpdateGoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Wireless Mouse",
            "description": "The name of the good"
          },
          "description": {
            "type": "string",
            "example": "A highly responsive wireless mouse.",
            "description": "The description of the good"
          },
          "quantity": {
            "type": "number",
            "example": 100,
            "description": "The quantity of the good available"
          },
          "price": {
            "type": "number",
            "example": 25.99,
            "description": "The price of the good"
          }
        }
      }
    }
  }
}